Solve part 2.
This has the bluntest possible solution to part 2. I was pretty
sure this would work because it's day 2, but I'd like to come
back and do something better.

Basically you want to run the program backwards and say, okay,
in order to have the target number in slot 0 at this final
instruction, what does the state of the machine need to be
before this instruction runs? Then just keep doing that backwards
and you should get the input out. I'm not 100% confident in that
because the machine can overwrite its own instructions so that
going backwards might make you look at a different machine that
going forwards, but it seems right.