Partial implementation of day 11.
The problem is pretty cagey about whether the brain's i/o
instructions will always happen in lockstep, so I'm not sure I
can rely on 2 values being available to read when the brain
blocks.  I might as well try and see how it goes, but I might
have to add a peek() method to intcode machines that lets me
see the available output without consuming it.